pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "computer-control-framework"
version = "0.1.0"
description = "跨平台电脑控制框架 - 为 AI Agent 设计的纯控制层"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Computer Control Framework Contributors"}
]
keywords = [
    "ai-agent",
    "computer-control",
    "automation",
    "screen-control",
    "ui-automation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pillow>=9.0.0",
    "pyautogui>=0.9.53",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Quartz>=9.0",
    "pyobjc-framework-ApplicationServices>=9.0",
]
ocr = [
    "easyocr>=1.7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/RuthlessXdream/computer-control-framework"
Documentation = "https://github.com/RuthlessXdream/computer-control-framework#readme"
Repository = "https://github.com/RuthlessXdream/computer-control-framework"
Issues = "https://github.com/RuthlessXdream/computer-control-framework/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

# ==================== Pytest ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow",
    "requires_permission: marks tests that require system permissions",
]

# ==================== Ruff (Linter) ====================
[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["src"]

# ==================== MyPy ====================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
